+++ controllers/authController.js
@@ -1,7 +1,7 @@
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcryptjs");
 const jwtDecode = require("jwt-decode");
-const { models } = require("../db/models");
+const { User } = require("../models");
 
 const createToken = (user) => {
   // Sign the JWT
@@ -25,7 +25,7 @@ const verifyPassword = (passwordAttempt, hashedPassword) => {
 const authPerson = async (req, res) => {
   try {
     const { email, password } = req.body;
-    const person = await models.User.find({ email }).exec();
+    const person = await User.find({ email });
 
     if (!person) {
       console.log("authPerson");
